update chat api
