+++ packages/app-copywriter/client/index.ts
@@ -1,4 +1,5 @@
 import Axios from "axios";
+import type { Message } from "../core/conversation";
 
 const client = Axios.create({
   baseURL: "/api",
@@ -14,9 +15,28 @@ interface TextPrompt {
   q: string;
 }
 
-export async function prompt(p: TextPrompt): Promise<TextResponse> {
+export async function copies(p: TextPrompt): Promise<TextResponse> {
   const { data } = await client.get<TextResponse>("/generative/copies", {
     params: p,
+  });
+
+  return data;
+}
+
+interface ChatPrompt {
+  content: string;
+  history: Message[];
+}
+
+interface ChatResponse {
+  q: string;
+  response: string;
+  model: string;
+}
+
+export async function chat(p: ChatPrompt): Promise<ChatResponse> {
+  const { data } = await client.post<ChatResponse>("/generative/chat", {
+    data: p,
   });
 
   return data;
